+++ CHANGELOG.md
@@ -5,10 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
 --------------------
+##### Added
+- In pages mode, the current page will be validated and needs to pass before going to the Next page.
+
 ##### Fixed
 - In pages mode, if the page contains a single question, this question does not get focus.
 - In pages mode, when the form is inside a scrollable container, it doesn't scroll to first question.
-
 
 [4.3.0] - 2015-10-08
 --------------------